+++ tests/test/core/TransactionTestUtils.h
@@ -54,7 +54,7 @@ namespace catapult { namespace test {
 	ConstTransactions MakeConst(const MutableTransactions& transactions);
 
 	/// Generates a random transaction with size \a entitySize.
-	std::unique_ptr<model::Transaction> GenerateRandomTransactionWithSize(size_t entitySize);
+	std::unique_ptr<model::Transaction> GenerateRandomTransactionWithSize(uint32_t entitySize);
 
 	/// Generates a transaction with \a deadline.
 	std::unique_ptr<model::Transaction> GenerateTransactionWithDeadline(Timestamp deadline);
